[CLN] Refactor `MergeMetadataResultsOperator`
